backend: refactor: remove firstTimeScraping.js
