+++ backend/test/twitterUserObjectScraping.test.js
@@ -2,14 +2,14 @@ const waitForExpect = require("wait-for-expect")
 const { executeCount } = require("../src/db/pg")
 const { setupPgBossQueue } = require("../src/db/pgBoss")
 const {
-  addFirstTimeTwitterUserObjectScrapingJobs,
+  addTwitterUserObjectScrapingJobs,
 } = require("../src/twitterUserObjectScraping")
-const { fillSampleOrgData } = require("./prepareDb")
+const { makeSampleOrgs } = require("./prepareDb")
 
 test("addFirstTimeTwitterUserObjectScrapingJobs", async () => {
-  await fillSampleOrgData()
   const pgBossQueue = await setupPgBossQueue()
-  await addFirstTimeTwitterUserObjectScrapingJobs(pgBossQueue)
+  const orgs = await makeSampleOrgs()
+  await addTwitterUserObjectScrapingJobs(pgBossQueue, orgs)
   await waitForExpect(
     async () => {
       const numScrapingResults = await executeCount("scraping_results")
